module4/char_recognition/requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
